[add]マウス入力・スキニングアニメーション・オーディオ追加
- Input: マウス対応（押下/トリガー/リリース、座標・移動量・ホイール、
  絶対/相対モード切替、カーソル表示制御）、GetKeyRelease追加。
  WndProcからInput::ProcessMessageでマウスメッセージをフック
- AnimationModel: assimpによるFBXスキニングモデル読み込み・再生を追加
  （assimpヘッダ、Akaiモデル/Idle/Runアニメーションを同梱）
- Audio: BGM/SEアセット追加、Playerにジャンプ音を実装
- Manager: ImGuiのNewFrameをDrawからUpdate先頭へ移動
- TitleScene: ロゴ・スタートボタンをメンバ化

+++ Object/Camera.cpp
@@ -36,7 +36,7 @@ void CAMERA::Update()
 
 	m_Target = playerPos;
 
-	m_Position = m_Target + Vector3(-sinf(m_Rotation.y) * 10.0f, 5.0f, -cos(m_Rotation.y) * 10.0f);
+	m_Position = m_Target + Vector3(-sinf(m_Rotation.y) * 10.0f, 5.0f, -cosf(m_Rotation.y) * 10.0f);
 }
 
 void CAMERA::Draw()
